requirements.txt
aiohttp
aiosignal
alpaca-trade-api
alpha-vantage
annotated-types
anthropic
anyio
attrs
certifi
charset-normalizer
colorama
deprecation
distro
dnspython
filelock
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
idna
jiter
msgpack
multidict
numpy
openai
packaging
pandas
pydantic
pydantic_core
pymongo
python-dateutil
python-dotenv
pytz
PyYAML
requests
scipy
six
sniffio
tokenizers
tqdm
typing_extensions
tzdata
urllib3
websocket-client
websockets
yarl
